% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.R
\name{bold}
\alias{bold}
\title{Bold}
\usage{
bold(table, pattern = "*")
}
\arguments{
\item{table}{data.frame. Table on which to highlight cell in bold.}

\item{pattern}{char. Pattern to be detected to highlight cells.}
}
\value{
data.frame
}
\description{
Render bold cell in Rmarkdown depending on a pattern (to put in blod
significant results for instance)
}
\examples{
# See draft.Rmd

}

