% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT.R
\name{CWT}
\alias{CWT}
\title{Community weighted trait}
\usage{
CWT(Trees, com, traits = c("LES", "SLA", "WES", "LDMC", "WD", "Thick", "LA"),
  metric = c("mean", "variance"), weights = c("abundance",
  "presence-absence", "basal area"), env = c("Slope", "Curvature", "Wetness",
  "SW", "Canopy", "BA"))
}
\arguments{
\item{Trees}{data.frame. Trees data frame}

\item{com}{SpatialGridDataFrame. Communities with their environment (abiotic
and biotic)}

\item{traits}{char. selected traits}

\item{metric}{char. selected metric to be computed on community}

\item{weights}{char. selected weights}

\item{env}{char. selected environmental variables to be kept}
}
\value{
a list with a community grid data frame for each weighting methods
  containing the environmental values, and the community weighted traits by
  defined metrics
}
\description{
Compute a community weighted metric (mean or variance) with traits
}
\examples{

}

