% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata}
\alias{extdata}
\title{External data}
\description{
To work, the package need external data. External data are raw files dataset
or raster needed for the analysis and loaded with gen functions.
}
\section{Files}{
 \describe{
\item{Canopy.tif}{Lidar canopy model}
\item{Coord_trees_allPlot.csv}{DB extraction for tree coordinates}
\item{data.csv}{PFT measurements}
\item{DEM.tif}{Digital elevation model}
\item{Girth_10Ha.csv}{DB extraction for 10 ha plots girth}
\item{Girth_5Ha.csv}{DB extraction for 5 ha plots girth}
\item{Height_5Ha.csv}{DB extraction for 5 ha plots height}
\item{Rivers.tif}{Distance to the nearest potential river model}
\item{Sp_trees_allPlot.csv}{DB extraction for tree species}
\item{Species.txt}{DB extraction for species data}
}
}

