% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genEnv.R
\name{genEnv}
\alias{genEnv}
\title{Environmental data generation}
\usage{
genEnv(area = data.frame(x = c(571675.8, 572011.3), y = c(1385704.7,
  1386010.8), row.names = c("min", "max")))
}
\arguments{
\item{area}{data.frame. Corners of studied area}
}
\value{
Environmental data set
}
\description{
Generates environmental data set from external data extracted from the database
}
\examples{
summary(genEnv())

}

