% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPFT.R
\name{genPFT}
\alias{genPFT}
\title{PFT data generation}
\usage{
genPFT()
}
\value{
PFT data set
}
\description{
Generates PFT data set from external data collected on the field
}
\examples{
summary(genPFT())

}

