% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSpecies.R
\name{genSpecies}
\alias{genSpecies}
\title{Species data generation}
\usage{
genSpecies()
}
\value{
Species data set
}
\description{
Generates Species data set from external data extracted from the database
}
\examples{
summary(genSpecies())

}

