% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTrees.R
\name{genTrees}
\alias{genTrees}
\title{Trees data generation}
\usage{
genTrees()
}
\value{
Trees data set
}
\description{
Generates Trees data set from external data extracted from the database
}
\examples{
summary(genTrees())

}

