% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{null_model}
\alias{null_model}
\title{Null model}
\usage{
null_model(com, Trees, PFT, formulas, SpCode, traits = c("LES", "SLA", "WES",
  "WD", "LDMC", "Thick", "LA"), env = c("Slope", "Curvature", "Wetness",
  "SouthWesterness", "Canopy", "BA"), type = c("CWM", "CWV"), n = 999,
  time = T, verbose = T)
}
\arguments{
\item{com}{list. List of communities}

\item{PFT}{data.frame. PFT values by species}

\item{formulas}{list. Model formulas}

\item{SpCode}{dta.frame. Link between species nd they're code}

\item{traits}{char. Choosen traits and strategy axes}

\item{env}{char. Choosen environmental variables}

\item{type}{char. Communti weighted trait metric}

\item{n}{int. number of repetitions}

\item{time}{logical. Use a timer}

\item{verbose}{logical. Use a counter}
}
\value{
List of communities null models
}
\description{
Function to compute null models
}
\examples{

}

