% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFTrel.R
\name{PFTrel}
\alias{PFTrel}
\title{PFT relation}
\usage{
PFTrel(PFT1, PFT2, data, captions, ...)
}
\arguments{
\item{PFT1}{char. Name of the first PFT}

\item{PFT2}{char. Name of the second PFT}

\item{data}{data.frame. PFT data set}

\item{captions}{vector. Captions for PFT}

\item{...}{other argument pass to plot}
}
\value{
Graphs of relation between the two PFT
}
\description{
PFT relation
}
\examples{
data(PFT)
data(captions)
PFTrel('SLA', 'WD', PFT, captions)

}

