% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeny.R
\name{phylogeny}
\alias{phylogeny}
\title{Phylogeny data}
\description{
To work, the package need external data from the Appendix of Qian, H. and Y.
Jin (2015). "An updated megaphylogeny of plants, a tool for generating plant
phylogenies, and an analysis of phylogenetic community structure." Journal of
Plant Ecology.
}

