% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.null_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot null model}
\usage{
plot_model(model, formula = NULL, index = NULL)
}
\arguments{
\item{model}{null_model list. result from null_model function (could do an object)}

\item{formula}{char. formula of the model to be plotted in the object}

\item{model}{int. indice of the model to be plotted in the object}
}
\value{
nothing but a plot of the result
}
\description{
Function to plot null model and get null model result
}
\examples{
# No example yet see Uppangala_PFT vignette

}

