% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrats.R
\name{quadrats}
\alias{quadrats}
\title{Quadrats}
\usage{
quadrats(cs = c(20, 20),
  proj = "+proj=utm +zone=43 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0",
  area = data.frame(x = c(571675.8, 572011.3), y = c(1385704.7, 1386010.8),
  row.names = c("min", "max")))
}
\arguments{
\item{cs}{int. Cell size (x*y quadrats)}

\item{proj}{CRS. Gird CRS}

\item{area}{data.frame. Corners of studied area}
}
\value{
List : com int. Vector with quadrat number & grd SpatialGridDataFrame. Quadrats'grid
}
\description{
Quadrats
}
\examples{
data(Trees)
quadrats(Trees$X, Trees$Y, n = 20)

}

