% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars}
\alias{stars}
\title{Stars}
\usage{
stars(val)
}
\arguments{
\item{val}{num. p-value to convert in stars}
}
\value{
char. Number of stars corresponding to the p-value
}
\description{
Stars
}
\examples{
stars(0.049)

}

