% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_classes.R
\name{topo_classes}
\alias{topo_classes}
\title{Topographical classes}
\usage{
topo_classes(Slope, Curvature, Aspect, degree = c(20, 30), dir = 225)
}
\arguments{
\item{Slope}{num. Slope model values}

\item{Curvature}{num. Curvature model values}

\item{degree}{num. Discriminant slope, autoset with NULL}

\item{dir}{num. Aspect direction to classify aspect in two categories (default SW for monsoon wind)}

\item{DEM}{num. Digital elevation model values}

\item{elev}{num. Discriminant elevation, autoset with NULL}
}
\value{
char. Vector of topographical classes
}
\description{
Topographical classes
}
\examples{
data("Env")
data("Trees")
topo_classes(Trees$X, Trees$Y, Env$DEM$Ground, Env$DEM$Slope, Env$DEM$Curvature)

}

