% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_deriv.R
\name{DEMderiv}
\alias{DEMderiv}
\title{Create DEM derivative}
\usage{
DEMderiv(DEM, choices = c("slope", "curvature", "plancurvature",
  "profcurvature", "aspect", "wetness", "flowdir", "hillshade"),
  minslope = 0.5, workspace = NULL,
  path = "C:/Program Files/QGIS Chugiak/apps/saga",
  method = "poly2zevenbergen")
}
\arguments{
\item{DEM}{raster. digital elevation model}

\item{choices}{char. choices of DEM derivative to compute}

\item{minslope}{num. for minslope, see help(rsaga.fill.sinks)}

\item{workspace}{char. current workspace}

\item{path}{char. path is the folder of the SAGA software}

\item{method}{char. for methods see help(rsaga.curvature)}
}
\value{
Raster stack with chosen DEM derivative
}
\description{
Create DEM derivative
}
\examples{
DEM <- genEnv()[[1]]
DEMderiv(DEM)

}
\author{
From Maxime Rejou-Mechain
}

