% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{null_model}
\alias{null_model}
\title{Null model}
\usage{
null_model(formula = "WD ~ topo", test = "Anova Fvalue", PFT_sp, Trees, com,
  SpCode, tcol = NULL, cwmcol = c(1:1), ccol = c(1:12), n = 1000,
  plot = F, count = F, time = T)
}
\arguments{
\item{formula}{char. test formula}

\item{test}{char. name of the used test (Anova Fvalue or lm R2)}

\item{PFT_sp}{data.frame. PFT values by species}

\item{Trees}{data.frame. trees table}

\item{com}{SpatialGridDataFrame. communities}

\item{tcol}{int. column to keep in trees table}

\item{ccol}{int. communities column to keep}

\item{n}{int. number of repetitions}

\item{plot}{logical. do all null communitier need to be plot}

\item{count}{logical. do the number of the realized null model need to be counted}

\item{time}{logical. do the time of the operation need to be measured}
}
\value{
a list with the formula, the test, the null values, and the community value
}
\description{
Function to compute null model
}
\examples{
# No example yet see Uppangala_PFT vignette

}

