% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.null_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot null model}
\usage{
plot_model(model)
}
\arguments{
\item{model}{null_model list. result from null_model function (could do an object)}
}
\value{
nothing but a plot of the result
}
\description{
Function to plot null model and get null model result
}
\examples{
# No example yet see Uppangala_PFT vignette

}

