% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_table.R
\name{regression_table}
\alias{regression_table}
\title{Regression table}
\usage{
regression_table(LM, reg, pval, ppval)
}
\arguments{
\item{LM}{list. List of tested regressions}

\item{reg}{char. Initial regression formula}

\item{model}{list. List of null models issued from null_model function}

\item{select}{char. subset of models to summarize}
}
\value{
A table summarizing all the regression
}
\description{
Regression table
}
\examples{
# No example yet

}

